add_JWT_CORS

+++ chapter7_security/planner/routes/events.py
@@ -1,9 +1,10 @@
-from fastapi import APIRouter, HTTPException, status
+from fastapi import APIRouter, Depends, HTTPException, status
 from beanie import PydanticObjectId
 from database.connection import Database
 from typing import List
 
 from models.events import Event, EventUpdate
+from auth.authenticate import authenticate
 
 
 event_database = Database(Event)
@@ -34,7 +35,8 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 
 @event_router.post("/new")
-async def create_event(body: Event) -> dict:
+async def create_event(body: Event, user: str = Depends(authenticate)) -> dict:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully."
@@ -42,24 +44,37 @@ async def create_event(body: Event) -> dict:
 
 
 @event_router.put("/{id}", response_model=Event)
-async def update_event(id: PydanticObjectId, body: EventUpdate) -> Event:
-    update_event = await event_database.update(id, body)
-    if not update_event:
+async def update_event(id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)) -> Event:
+    event = await event_database.get(id)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
+    updated_event = await event_database.update(id, body)
+    if not updated_event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="Event with supplied ID does not exist."
+            detail="Event with supplied ID does not exist"
         )
-    return update_event
+    return updated_event
 
 
 @event_router.delete("/{id}")
-async def delete_event(id: PydanticObjectId) -> dict:
-    event = await event_database.delete(id)
+async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
+    event = await event_database.get(id)
     if not event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="Event with supplied ID does not exist."
+            detail="Event not found"
         )
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed"
+        )
+    event = await event_database.delete(id)
+
     return {
-            "message": "Event deleted successfully."
-        }
+        "message": "Event deleted successfully."
+    }
